処理のリファクタリング

+++ src/main/java/com/example/create_db_file/schedule/CleanTemporalFile.java
@@ -20,36 +20,44 @@ public class CleanTemporalFile {
 
     private final ResourceLoader resourceLoader;
 
-    //    @Scheduled(fixedDelay = 1000)
-
     /**
      * 10分起きに一時ファイルディレクトリに格納されているファイルを精査し
      * 保存してから10分以上経過しているファイルを削除するメソッド
      */
     @Scheduled(fixedDelay = 1000 * 60 * 10)
+//    @Scheduled(fixedDelay = 1000 * 60)
     public void sampleTask(){
         log.info("hello scheduled task!!");
         Resource resource = resourceLoader.getResource("classpath:temporal");
         DateTimeFormatter formatter = DateTimeFormatter.ofPattern("yyyyMMddHHmm");
         String formatNowDate = LocalDateTime.now().format(formatter);
+
         try {
+
+            long fileCount = Files.walk(resource.getFile().toPath()).count();
+            log.info("現在の格納ファイル数: {}", fileCount);
             Files.walk(resource.getFile().toPath()).filter(path -> !Files.isDirectory(path))
                     .forEach(path -> {
-                        log.info(path.toString());
-                        log.info(path.getFileName().toString());
-                        log.info(StringUtils.substring(path.getFileName().toString(),0, 12));
-                        String temporalFileInfo = StringUtils.substring(path.getFileName().toString(),0, 12);
-                        long timeLag = Long.valueOf(formatNowDate) - Long.valueOf(temporalFileInfo);
-                        if(timeLag >= 10){
-                            try {
-                                Files.delete(path);
-                            } catch (IOException e) {
-                                e.printStackTrace();
+
+                        if(path.getFileName().toString().length() > 12 && path.getFileName().toString().substring(0, 12).matches("[0-9]*")){
+                            String temporalFileInfo = StringUtils.substring(path.getFileName().toString(),0, 12);
+                            long timeLag = Long.parseLong(formatNowDate) - Long.parseLong(temporalFileInfo);
+                            // if(timeLag >= 1)
+                            if(timeLag >= 10){
+                                try {
+                                    Files.delete(path);
+                                    log.info("ファイル名: {} を削除しました", path.getFileName());
+                                } catch (IOException e) {
+                                    log.info("ファイル削除処理中にエラー発生",e);
+                                }
                             }
+
+                        } else {
+                            log.info("削除対象ファイルでは無いため無視されました: 無視されたファイル -> {}", path.getFileName());
                         }
                     });
         } catch (IOException e) {
-            e.printStackTrace();
+            log.info("ファイル削除処理中にエラー発生",e);
         }
     }
 }
